+++ packages/transaction-pay-controller/src/utils/strategy.ts
@@ -1,4 +1,5 @@
 import { TransactionPayStrategy } from '../constants';
+import { AcrossStrategy } from '../strategy/across/AcrossStrategy';
 import { BridgeStrategy } from '../strategy/bridge/BridgeStrategy';
 import { RelayStrategy } from '../strategy/relay/RelayStrategy';
 import { TestStrategy } from '../strategy/test/TestStrategy';
@@ -19,6 +20,9 @@ export function getStrategyByName(
   strategyName: TransactionPayStrategy,
 ): PayStrategy<unknown> {
   switch (strategyName) {
+    case TransactionPayStrategy.Across:
+      return new AcrossStrategy() as never;
+
     case TransactionPayStrategy.Bridge:
       return new BridgeStrategy() as never;